new_mem_api: move some function to xio_mem
This commit moves some memory related functions
from xio_rdma_verbs and xio_usr_transport to xio_mem.
This is imporves code readabilty by grouping similar
functions that are effected by the same IFDEF to be
under the same file.
