+++ src/common/xio_transport.h
@@ -38,6 +38,8 @@
 #ifndef XIO_TRANSPORT_H
 #define XIO_TRANSPORT_H
 
+#include "xio_observer.h"
+
 /*---------------------------------------------------------------------------*/
 /* forward declarations	                                                     */
 /*---------------------------------------------------------------------------*/
