+++ src/usr/transport/xio_usr_transport.c
@@ -54,87 +54,6 @@
 #include "xio_workqueue.h"
 #include "xio_context.h"
 
-#ifndef HAVE_INFINIBAND_VERBS_H
-
-/*---------------------------------------------------------------------------*/
-/* xio_mem_register							     */
-/*---------------------------------------------------------------------------*/
-int xio_mem_register(void *addr, size_t length, struct xio_reg_mem *reg_mem)
-{
-	static struct xio_mr dummy_mr;
-
-	if (!addr || !reg_mem) {
-		xio_set_error(EINVAL);
-		return -1;
-	}
-
-	reg_mem->addr = addr;
-	reg_mem->length = length;
-	reg_mem->mr = &dummy_mr;
-
-	return 0;
-}
-
-/*---------------------------------------------------------------------------*/
-/* xio_mem_dereg							     */
-/*---------------------------------------------------------------------------*/
-int xio_mem_dereg(struct xio_reg_mem *reg_mem)
-{
-	reg_mem->mr = NULL;
-	return 0;
-}
-
-/*---------------------------------------------------------------------------*/
-/* xio_mem_alloc							     */
-/*---------------------------------------------------------------------------*/
-int xio_mem_alloc(size_t length, struct xio_reg_mem *reg_mem)
-{
-	size_t			real_size;
-	int			alloced = 0;
-
-	real_size = ALIGN(length, page_size);
-	reg_mem->addr = umemalign(page_size, real_size);
-	if (!reg_mem->addr) {
-		ERROR_LOG("xio_memalign failed. sz:%zu\n", real_size);
-		goto cleanup;
-	}
-	/*memset(reg_mem->addr, 0, real_size);*/
-	alloced = 1;
-
-	xio_mem_register(reg_mem->addr, length, reg_mem);
-	if (!reg_mem->mr) {
-		ERROR_LOG("xio_reg_mr failed. addr:%p, length:%d\n",
-			  reg_mem->addr, length, access);
-
-		goto cleanup1;
-	}
-	reg_mem->length = length;
-
-	return 0;
-
-cleanup1:
-	if (alloced)
-		ufree(reg_mem->addr);
-cleanup:
-	return -1;
-}
-
-/*---------------------------------------------------------------------------*/
-/* xio_mem_free								     */
-/*---------------------------------------------------------------------------*/
-int xio_mem_free(struct xio_reg_mem *reg_mem)
-{
-	int			retval = 0;
-
-	if (reg_mem->addr)
-		ufree(reg_mem->addr);
-
-	retval = xio_mem_dereg(reg_mem);
-
-	return retval;
-}
-
-#endif /*HAVE_INFINIBAND_VERBS_H*/
 
 /*---------------------------------------------------------------------------*/
 /* xio_transport_mempool_get						     */
